move BlueEntity and rename object to value
