+++ BlueDBOnDisk/src/main/java/io/bluedb/disk/collection/task/UpdateMultipleTask.java
@@ -10,7 +10,7 @@ import io.bluedb.api.exceptions.BlueDbException;
 import io.bluedb.disk.collection.BlueCollectionImpl;
 import io.bluedb.disk.recovery.PendingChange;
 import io.bluedb.disk.recovery.RecoveryManager;
-import io.bluedb.disk.segment.BlueEntity;
+import io.bluedb.disk.serialization.BlueEntity;
 import io.bluedb.disk.serialization.BlueSerializer;
 
 public class UpdateMultipleTask<T extends Serializable> extends QueryTask {
