+++ BlueDBOnDisk/src/main/java/io/bluedb/disk/segment/Segment.java
@@ -10,6 +10,7 @@ import io.bluedb.api.exceptions.BlueDbException;
 import io.bluedb.api.keys.BlueKey;
 import io.bluedb.api.keys.TimeFrameKey;
 import io.bluedb.disk.file.FileManager;
+import io.bluedb.disk.serialization.BlueEntity;
 
 public class Segment <T extends Serializable> {
 
@@ -64,7 +65,7 @@ public class Segment <T extends Serializable> {
 		List<T> results = new ArrayList<>();
 		for (File file: filesInFolder) {
 			for (BlueEntity<T> entity: fetch(file)) {
-				results.add(entity.getObject());
+				results.add(entity.getValue());
 			}
 		}
 		return results;
@@ -130,7 +131,7 @@ public class Segment <T extends Serializable> {
 			BlueEntity<T> entity = entities.get(i);
 			if (entity.getKey().equals(key)) {
 				entities.remove(i);
-				return entity.getObject();
+				return entity.getValue();
 			}
 		}
 		return null;
@@ -143,7 +144,7 @@ public class Segment <T extends Serializable> {
 	protected static <T extends Serializable> T get(BlueKey key, List<BlueEntity<T>> entities) {
 		for (BlueEntity<T> entity: entities) {
 			if (entity.getKey().equals(key)) {
-				return entity.getObject();
+				return entity.getValue();
 			}
 		}
 		return null;
